[ref #190] refactor
